# application/apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOFA-Guided Oncology Research Copilot API",
    "version": "0.2.0"
  },
  "paths": {
    "/analysis/{run_id}": {
      "get": {
        "summary": "Get Analysis",
        "operationId": "get_analysis_analysis__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/{run_id}/progress": {
      "get": {
        "summary": "Progress",
        "operationId": "progress_analysis__run_id__progress_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/demo/{patient_id}": {
      "post": {
        "summary": "Demo",
        "operationId": "demo_analysis_demo__patient_id__post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSubmitAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisProgressOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/AnalysisStageOut"
            },
            "type": "array",
            "title": "Stages",
            "default": []
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "AnalysisProgressOut"
      },
      "AnalysisResultOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "patient_label": {
            "type": "string",
            "title": "Patient Label"
          },
          "patient_metadata": {
            "$ref": "#/components/schemas/PatientMetadata"
          },
          "administered_regimen": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Administered Regimen"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          },
          "signature_params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Params"
          },
          "cluster_prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusterPredictionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "rna_projection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RnaProjectionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "cluster_signature": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignaturePanelOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "residual_signature": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignaturePanelOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "overlap_nominations": {
            "items": {
              "$ref": "#/components/schemas/OverlapNominationOut"
            },
            "type": "array",
            "title": "Overlap Nominations",
            "default": []
          },
          "overlap_exploratory": {
            "items": {
              "$ref": "#/components/schemas/OverlapNominationOut"
            },
            "type": "array",
            "title": "Overlap Exploratory",
            "default": []
          },
          "overlap_technical_excluded": {
            "items": {
              "$ref": "#/components/schemas/OverlapNominationOut"
            },
            "type": "array",
            "title": "Overlap Technical Excluded",
            "default": []
          },
          "display_gate_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Gate Summary"
          },
          "compound_registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compound Registry Version"
          },
          "analysis_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Summary"
          },
          "clinical_comparators": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Clinical Comparators",
            "default": []
          },
          "predictor_single_drugs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Predictor Single Drugs",
            "default": []
          },
          "predictor_combinations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Predictor Combinations",
            "default": []
          },
          "predictor_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predictor Summary"
          },
          "near_consensus": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Near Consensus",
            "default": []
          },
          "overlap_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap Summary"
          },
          "almanac_combinations": {
            "items": {
              "$ref": "#/components/schemas/AlmanacCombinationOut"
            },
            "type": "array",
            "title": "Almanac Combinations",
            "default": []
          },
          "list1_drugs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "List1 Drugs",
            "default": []
          },
          "list2_drugs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "List2 Drugs",
            "default": []
          },
          "administered_regimen_pcr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupportedPcrResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "top_candidate_drugs": {
            "items": {
              "$ref": "#/components/schemas/DrugCandidate"
            },
            "type": "array",
            "title": "Top Candidate Drugs",
            "default": []
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations",
            "default": []
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningOut"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "prototype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrototypePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "s4_ships": {
            "type": "boolean",
            "title": "S4 Ships",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "created_at",
          "patient_label",
          "patient_metadata",
          "administered_regimen"
        ],
        "title": "AnalysisResultOut"
      },
...
      "AnalysisSubmitAck": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "poll_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll Url"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "AnalysisSubmitAck"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientMetadata": {
        "properties": {
          "age_at_diagnosis": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age At Diagnosis"
          },
          "er_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Er Status"
          },
          "her2_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Her2 Status"
          },
          "pr_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Status"
          },
          "claudin_subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claudin Subtype"
          },
          "histological_subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Histological Subtype"
          },
          "lymph_nodes_positive": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lymph Nodes Positive"
          },
          "menopausal_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Menopausal State"
          },
          "nottingham_prognostic_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nottingham Prognostic Index"
          },
          "tumor_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tumor Stage"
          },
          "tumor_grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tumor Grade"
          },
          "tumor_size_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tumor Size Mm"
          },
          "ecog_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ecog Status"
          },
          "prior_therapy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prior Therapy"
          },
          "organ_function": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organ Function"
          },
          "location": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "field_provenance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Provenance"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "PatientMetadata",
        "description": "Oncology metadata. Extra demo-enriched fields are allowed and preserved."
      },
...
      "SupportedPcrResult": {
        "properties": {
          "applicability_gate": {
            "$ref": "#/components/schemas/ApplicabilityGate"
          },
          "mofa_regimen_reversal_percentile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mofa Regimen Reversal Percentile"
          },
          "mofa_regimen_reversal_note": {
            "type": "string",
            "title": "Mofa Regimen Reversal Note"
          },
          "pcr_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pcr Probability"
          },
          "q2_regimen_score": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q2 Regimen Score"
          }
        },
        "type": "object",
        "required": [
          "applicability_gate",
          "mofa_regimen_reversal_percentile",
          "mofa_regimen_reversal_note",
          "pcr_probability",
          "q2_regimen_score"
        ],
        "title": "SupportedPcrResult"
      },
...
    }
  }
}